+++ src/service-worker.ts
@@ -115,14 +115,18 @@ export async function generateServiceWorker(options: AddServiceWorkerOptions):
   // swPrecache will determine the right urls by stripping buildRoot.
   // NOTE:(usergenic) sw-precache generate() apparently replaces the
   // prefix on an already posixified version of the path on win32.
-  swPrecacheConfig.stripPrefix = posixifyPath(buildRoot);
+  //
+  // We include a trailing slash in `stripPrefix` so that we remove leading
+  // slashes on the pre-cache asset URLs, hence producing relative URLs
+  // instead of absolute. We want relative URLs for builds mounted at non-root
+  // paths. Note that service worker fetches are relative to its own URL.
+  swPrecacheConfig.stripPrefix = addTrailingSlash(posixifyPath(buildRoot));
 
   if (options.basePath) {
+    // TODO Drop this feature once CLI doesn't depend on it.
     let replacePrefix = options.basePath;
-    // Avoid a "/foo//bar" situation.
-    if (replacePrefix.endsWith('/') &&
-        !swPrecacheConfig.stripPrefix.endsWith('/')) {
-      replacePrefix = replacePrefix.slice(0, -1);
+    if (!replacePrefix.endsWith('/')) {
+      replacePrefix = replacePrefix + '/';
     }
     if (swPrecacheConfig.replacePrefix) {
       console.info(
@@ -173,4 +177,8 @@ export function addServiceWorker(options: AddServiceWorkerOptions):
       });
     });
   });
+}
+
+function addTrailingSlash(s: string): string {
+  return s.endsWith('/') ? s : s + '/';
 }